make measurement files corruption safe.
at the cost of some i/o cost (one copy of the meas file at each checkpointing),
we can now assure that the results of the simulation are always in a valid state,
even if the program crashes while writing them out. At least in theory. Let’s see
how it works.

We also changed to the std::filesystem API… meaning there could be bugs…
and loadleveller probably runs on Windows now!