remove dependence on the working directory
instead of $(pwd)/JOBNAME.data, the job director is now the directory containing the parameter file.
This will not change the behavior of the loadl script, which changes directory and creates the parameters file at
$(pwd)/JOBNAME.data/parameters.json

However, when you used loadleveller without loadl, there were subtle complications when you do not get the working directory hopping correct.
The new behavior is more predictable and helps with that. One slightly ugly artifact is that the .results.json file gets created in “the parent directory of the job directory” which can be really messy if you don’t follow the same convention loadl uses. But I’ll leave it like this for
the sake of simplicity.

For good measure, I have extended the use of std::filesystem::path in the code.