remove the old mersenne twister and intel_mkl PRNGs
I wrote a custom mt19937 implementation to replace it.
Using the available libraries (stl, intel_mkl) in production
code is not feasible because significant parts of them cannot
be inlined. Our implementation is (like the old one) header-only
and inlined completely.

It is also gives identical random sequences for a given (internal)
seed as the old MersenneTwister.h and std::mt19937. As it is checkpoint
compatible with MersenneTwister.h, it should not affect existing
runs.